refactor: rename some snapshot classes to fit the meaning, group them in one directory
