+++ src/MarginTrading.Backend.Core/Services/ISnapshotValidationService.cs
@@ -2,6 +2,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Threading.Tasks;
+
 using MarginTrading.Backend.Core.Snapshots;
 
 namespace MarginTrading.Backend.Core.Services
@@ -15,6 +16,6 @@ namespace MarginTrading.Backend.Core.Services
         /// Validates current trading state for inconsistent state.
         /// </summary>
         /// <returns>The validation result that contains the details of the mismatch between the current and expected status..</returns>
-        Task<SnapshotValidationResult> ValidateCurrentStateAsync();
+        Task<EnvironmentValidationResult> ValidateCurrentStateAsync();
     }
 }
