+++ src/MarginTrading.Backend.Services/Extensions/TradingEngineSnapshotExtensions.cs
@@ -13,6 +13,7 @@ using MarginTrading.Backend.Core;
 using MarginTrading.Backend.Core.Orders;
 using MarginTrading.Backend.Core.Snapshots;
 using MarginTrading.Backend.Core.Trading;
+using MarginTrading.Backend.Services.Snapshot;
 
 namespace MarginTrading.Backend.Services.Extensions
 {
