--- src/MarginTrading.Backend.Services/ISnapshotValidator.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-
-using MarginTrading.Backend.Core.Snapshots;
-
-namespace MarginTrading.Backend.Services;
-
-public interface ISnapshotValidator
-{
-    Task<SnapshotValidationResult> Validate(string correlationId);
-}
