+++ src/MarginTrading.Backend.Services/Modules/ManagersModule.cs
@@ -9,6 +9,7 @@ using MarginTrading.Backend.Services.Caches;
 using MarginTrading.Backend.Services.Infrastructure;
 using MarginTrading.Backend.Services.MatchingEngines;
 using MarginTrading.Backend.Services.Quotes;
+using MarginTrading.Backend.Services.Snapshot;
 using MarginTrading.Backend.Services.TradingConditions;
 
 namespace MarginTrading.Backend.Services.Modules
@@ -65,8 +66,8 @@ namespace MarginTrading.Backend.Services.Modules
                 .As<ISnapshotValidationService>()
                 .SingleInstance();
 
-            builder.RegisterType<SnapshotValidator>()
-                .As<ISnapshotValidator>()
+            builder.RegisterType<EnvironmentValidator>()
+                .As<IEnvironmentValidator>()
                 .SingleInstance();
 
             builder.RegisterType<DraftSnapshotWorkflowTracker>()
