+++ src/MarginTrading.Backend.Services/Snapshot/IEnvironmentValidator.cs
@@ -1,0 +1,15 @@
+using System.Threading.Tasks;
+
+using MarginTrading.Backend.Core.Snapshots;
+
+namespace MarginTrading.Backend.Services.Snapshot;
+
+public interface IEnvironmentValidator
+{
+    /// <summary>
+    /// Checks the environment consistency before creating the snapshot
+    /// </summary>
+    /// <param name="correlationId"></param>
+    /// <returns></returns>
+    Task<EnvironmentValidationResult> Validate(string correlationId);
+}
