+++ src/MarginTrading.Backend.Services/Snapshot/IFinalSnapshotCalculator.cs
@@ -3,10 +3,11 @@
 
 using System.Collections.Generic;
 using System.Threading.Tasks;
+
 using MarginTrading.Backend.Contracts.Prices;
 using MarginTrading.Backend.Core.Snapshots;
 
-namespace MarginTrading.Backend.Services
+namespace MarginTrading.Backend.Services.Snapshot
 {
     /// <summary>
     /// Trading snapshot recalculation from draft to final using closure prices
