+++ src/MarginTrading.Backend.Services/Snapshot/TradingEngineSnapshotLoggingBuilder.cs
@@ -10,10 +10,11 @@ using MarginTrading.Backend.Core;
 using MarginTrading.Backend.Core.Orders;
 using MarginTrading.Backend.Core.Snapshots;
 using MarginTrading.Backend.Core.Trading;
+using MarginTrading.Backend.Services.Infrastructure;
 
 using Microsoft.Extensions.Logging;
 
-namespace MarginTrading.Backend.Services.Infrastructure;
+namespace MarginTrading.Backend.Services.Snapshot;
 
 class TradingEngineSnapshotLoggingBuilder(
     TradingEngineSnapshotBuilder decoratee,
