+++ src/MarginTrading.Backend/Modules/BackendServicesModule.cs
@@ -23,6 +23,7 @@ using MarginTrading.Backend.Services.Quotes;
 using MarginTrading.Backend.Services.RabbitMq;
 using MarginTrading.Backend.Services.Services;
 using MarginTrading.Backend.Services.Settings;
+using MarginTrading.Backend.Services.Snapshot;
 using MarginTrading.Common.Services;
 
 namespace MarginTrading.Backend.Modules
