+++ src/MarginTrading.Backend/Startup.cs
@@ -47,6 +47,7 @@ using MarginTrading.Backend.Services.Modules;
 using MarginTrading.Backend.Services.Quotes;
 using MarginTrading.Backend.Services.Scheduling;
 using MarginTrading.Backend.Services.Settings;
+using MarginTrading.Backend.Services.Snapshot;
 using MarginTrading.Backend.Services.TradingConditions;
 using MarginTrading.Common.Extensions;
 using MarginTrading.Common.RabbitMq;
