+++ tests/MarginTradingTests/PositionsProviderTests.cs
@@ -9,6 +9,8 @@ using Autofac.Extras.Moq;
 using MarginTrading.Backend.Core.Orders;
 using MarginTrading.Backend.Services;
 using MarginTrading.Backend.Services.Services;
+using MarginTrading.Backend.Services.Snapshot;
+
 using Moq;
 using NUnit.Framework;
